Added new "simple" integration function.
Multiplication of opvar2d and dopvar2d objects requires a lot of integrals to be evaluated. However, all of these integrals have a simple 'pvar' or 'double' limit, rather than a complicated polynomial expression. The "int_simple" function is tailored to this particular case, aiming to reduce the computation time associated with operations on opvar2d and dopvar2d objects.