Update dopvar2d class functions.
Minor bugfixes and performance enhancements. 
The biggest change is the new "set" function. Dimensions of an opvar2d class object can now only be adjusted if all parameters along this dimension are zeroes (of a particular size). These zeroes will then be expanded/compressed to match the new desired dimensions. When adjusting var1 and var2 fields, the variables in all parameters of the operator are adjusted to match the new variable names. Finally, new parameters are now extensively checked to see if their dimensions and variables match the current dimensions and variables of the operator. If a particular dimension does not match, but all parameters along this dimension are zeroes, these parameters will be expanded/compressed to the new dimensions. Otherwise, an error will be produced. If the variables in the new parameter do not match current variables, but also do not conflict with variables in current parameters of the operator, the new variables will be used to determine var1 and var2, if possible. If the new variables clearly conflict with variables in other parameters, an error will be produced.
In addition, "mtimes" has been adjusted to use "int_simple", which is an alternative to the standard polynomial "int", tailored to a specific situation where the limits of the integral are "simple" expressions. The aim is to reduce the computation time associated with multiplying dopvar2d and opvar2d objects.