Update to support edge inputs.
Previously the function could only be applied with values specified as numeric values. Now if the user requests a variable to be evaluated at itself (i.e. invar=x and inval=x), the original operator is simply returned. This is to simplify evaluation of the operator at the edges, setting e.g. invar=[x;y] and inval=[x;0] or inval=[0;y] to evaluate the operator at the lower or left boundary.