Bugifx blkdiag opvar2d with polynomial.
The function should only allow "blkdiag" on a polynomial with an opvar2d object to be performed if the opvar2d object maps from and to a single function space (e.g. L2[x] --> L2[x,y] rather than R x L2[x] --> L2[x] x L2[x,y]), as otherwise it is unclear what function spaces the polynomial maps to/from. However, the function previously failed to properly check for this condition to be satisfied, resulting in an erroneous output. This has been fixed.