Upload initialization for new terms format.
Uploaded a function for the initialization of PDE structures in the new data format. Also included a new class object to represent the PDEs in the new terms format, though I'm not sure it contributes much. It will likely need to be moved to another folder, or just removed. It does however have a display function to actually show what PDE the user specifies, so that might be useful.

Lacking in the initialization function right now:
- Spatial variables must be specified for each component, or it will revert to a finite-dimensional component. I think it would be preferable that if a spatial domain is specified, but no spatial variables, we just use standard variables.
- Different terms concerning the same derivative of the same state at the same position are not currently added. I don't think this feature is important, but it might be nice to include at some point.