Update PDE initialization function.
The function previously required a spatial position to be specified in any term of the BCs. This has been adjusted, requiring no spatial position to be specified if a full integral is taken along any spatial direction, in which case the spatial position (if not specified) will default to the interior (s1=s1, s2=s2, etc.).