Bugfix dopvar display.
The display function for dpvar objects asks the user if they want to continue with the display if the object is found to be very large. This produced an error when trying to display parameters of dopvar objects. This has been resolved, not displaying parameters which the dpvar display function determines to be too large.