Add "@pde_struct/convert".
A convert function has been added to the "pde_struct" class functions, so that "convert(PDE)" can be called rather than having to call "convert_PIETOOLS_PDE(PDE)". This is to match the notation for conversion of "sys" class objects to PIEs, so that users can call the same function for both types of objects. An "initialize" function has also been added to the "pde_struct" class functions, so that "initialize(PDE)" can be called to initialize the PDE.