Update pde_struct files.
Added a file to merge variables in the PDE, to decrease the dimensionality if possible. Also working on an update to include delays, though some fixes are still in order. Finally, there was an issue with the initialize function where certain variables may end up getting swapped (unlikely in 1D or 2D, but more likely in higher dimensions). This issue should (should) be fixed, but will require some additional testing to be sure.