Bugfix PDE terms to PIE converter.
In specifying the contributions of integrals and boundary values of the PDE state to the outputs, the function checked whether a field "I" (specifying an integral) was present, assuming a boundary term only if this was not the case. However, an empty integral could also be specified, in which case the function would still erroneously assume the PDE state to be integrated. The function has been updated to also check if the field "I" is empty, assuming a boundary term only if no "I" is present, or "I" is empty.