Update lpi_eq functions.
In most applications, the operators which we set equal to zero will be symmetric, so that e.g. Pop.Q1=Pop.Q2'. In that case, in requiring Pop=0 it suffices to enforce only Pop.Q1=0 to insure that also Pop.Q2=0, thus allowing the complexity of the SDP to be reduced slightly. The updated functions add an optional argument to lpi_eq, so that when this argument is set to 'symmetric', the operator Pop is assumed to be symmetric, and only the "lower-triangular" parameters in this operator are set equal to zero.