Update figures and manual.
The figure for Demo 5 now corresponds to that using the executive rather than using the manual implementation, also in the manual.
The examples in Sections 7.3.1 and 7.3.2 of the manual have been updated to account for the fact that concatenation of opvars with matrices is no longer supported.
Section 7.6.2c has been updated to include the new functionality for ```closedLoopPIE```.
Section 10.3 has been updated to note that ambiguous ```opvar``` concatenations are no longer supported.
Some slight changes have been made to Sections 11.5-11.7 to incorporate the new ```closedLoopPIE``` function.