Update lpi_ineq_2d.m
Fix parameter checks.
Also include option to add further freedom in the positive opvar2d object, by adding separable operators. This option was previously commented, but can now be used by setting toggle2=2. This option is mainly for testing purposes, so "toggle2" is not an argument of the function, and is set to 0 by default.