Add declaration of PDE vars.
The function 'pde_var' allows a state, input, or output variable to be declared, of which the size, the spatial variables, and domain of the spatial variables can be set. Information on the variable will be stored in the field `PDE.x`, `PDE.y`, `PDE.z`, etc., depending on what type of variable is declared. The variable can then be used to construct free terms to be build a PDE, using command line operations. To this end, a conversion 'var2term' is also added to declare a single free term (in the field `PDE.free{1}.term{1}` that corresponds to just this variable. Finally, functions were added to check whether a PDE structure `PDE` corresponds to a free PDE variable, or a set of free terms.