Create size for pde_struct
Mainly for internal use. This checks the number of equations specified by the PDE structure (number of rows), and the maximal number of terms that appears in the equations (number of columns). Number of equations of a particular type (state, output, BC) can also be extracted using optional input arguments.