Add differentiation for PDE terms.
The function 'diff' allows the spatial derivative of a state variable to be taken. A temporal derivative can also be taken, using `t` as the variable. In this case, the resulting term can only be used in the left-hand side of the PDE. To distinguish such "left-hand side" terms, the function 'is_LHS_term' checks whether a given term corresponds to an output signal, or a temporal derivative of a state. This will be necessary for when we declare an actual equation using 'eq'.