Create subs for pde_struct.
Allows free PDE variables (states, inputs, outputs) to be evaluated at a particular positions, or their spatial variable to be substituted with another variable for e.g. integration. Note that this does not change the spatial variable linked to the PDE object, but merely declares a term wherein the old spatial variable of the PDE object is substituted by the new one.