Update subsref and subsasgn.
The updated subsref function allows term j in equation i of a PDE `PDE` to be retrieved as `PDE(i,j)`. Ideally, subsasgn would be updated so that the term (i,j) could be replaced as `PDE(i,j)=new_term`, but this has not happened. However, 'subsasgn' has been updated to allow the size, variables, and domain of a PDE variable (state, input, output) to be set as e.g. `x.size=2`, `x.vars=[ss1;ss2]`, and `x.dom=[0,1;1,2]`. The function 'subsref' has been updated to allow these field to be retrieved as well.