Update subs.m
Add support for substitution of input signals. At this time, initialization of PDEs with inputs evaluated anywhere is not supported, and therefore also no conversion to a PIE. But, partial integrals of inputs should be allowed, which requires substituting the variable with a dummy variable, and more general substitution may also come into play for forcing at the boundary for higher-dimensional PDEs.