Add statenames to pde_struct.
When calling "initialize", the IDs associated to different states, inputs, and outputs were previously overwritten with 1 through n_obj, where n_obj is the number of objects of the considered type. However, for PDEs constructed in the command line, this may lead to confusion, as the IDs are suddenly changed. The function `initialize` has now been updated to retain the original IDs associated to the objects, or generate a new ID if none has been specified. However, these IDs are now replaced when calling `reorder_comps`, for proper conversion to the PIE format.