Add "ambiguous" concatenations.
The latest versions of `vertcat` and `horzcat` of 'opvar' and 'opvar2d' objects do not allow for concatenations if the dimensions of the resulting operator are ambiguous. However, for the constructing of e.g. the Hinfty-norm LPI, if the input or output signals are infinite dimensional, this leads to errors. Thus, for now, 'legacy' versions of the concatenation have been added for at least the 'opvar2d' class. Something similar may need to be done for the 'opvar' class, though really, a more elegant solution should be found.