Update PIESIM_transform_to_solution_2D.m
Currently, PIESIM does not support forcing at the boundary in 2D, and so this forcing is set to zero. However, the dimensions of this zero array result in conflicts when adding this zero contribution to the numerical solutions. Since I do not know how exactly the boundary forcing is to be incorporated, I have for now added a check to not add the contribution of this forcing if it is zero, to avoid errors.