Bugfix non-polynomial disturbance.
The contribution of non-polynomial, spatially distributed disturbances are processed separately using `discretize_icf` and re-incorporated at the end of `discretize_all`. However, for PDEs involving multiple state variables, this non-polynomial contribution was added only to the differential equation for the first state variable. This has been resolved, now checking which state components each disturbance contributes to, and then adding the non-polynomial contribution only to the equations for those state variables.