Update PIESIM_input_check.m
There was a minor issue if initial conditions for ODEs were specified using an object of type `sym`, throwing an error even if this object corresponds to a constant value. This has been adjusted to throw an error only if the symbolic object indeed depends on some variable, and convert to type 'double' otherwise.