Update PIESIM_plot_solution_2D.m
There was a minor issue with the indexing of the exact state solutions in the case of coupled 1D and 2D PDEs in lines 166 (now 167) 210 (now 212) and 255 (now 258). In particular, in looping over the 1D and 2D state components for extraction of the exact solutions, the number of state components from previous loops was not taken into account, resulting in the same exact solution to be linked to different state components. This has been resolved, keeping track of the total number of state components using the added variable `ns_tot`.