Update diff.m
Fixed bugs related to differentiation of PDE state variables. In particular, an issue has been resolved with specifying multiple derivatives as `diff(diff(x1,s1),s2)`, which was caused by using incorrect brackets `()` rather than `{}`. In addition, derivatives of states with respect to variables on which they do not depend no longer return an error, but just return a 0 term.