Create mrdivide_poly.m
This function acts as `mrdivide` function for polynomials, taking input 'polynomial' objects `A` and `b`, and returning a 'polynomial' object 'x' that approximately solves `x*A=b`. If there exists a polynomial `x` for which this equation holds, then the function should be able to find this `x`. Otherwise, the approximation is not necessarily optimal in any sense, but a tolerance can be specified to enforce '||x*A-b||_{L2}<tol', increasing the maximal monomial degree of 'x' until this is satisfied. This function can be used to compute an approximate left-inverse of a polynomial `p` by calling `mrdivide_poly(1,p)`.