Update opvar2d inversion.
A function `is_separable` has been added to check if an 'opvar2d' object represents a separable 2D PI operator. This is the case only if any map L2[vars1]-->L2[vars2] acts as integral operator along vars1, and multiplier operator along vars2. That is, even if there is partial overlap in vars1 and vars2, the operator is not separable if it acts as multiplier operator along these shared variables. However, if vars1=vars2, then the operator may include both a multiplier term and full integral term.
The updated `inv` function checks if the input operator is separable.
If so, the inverse can be computed analytically, which is then done using the updated function `inv_opvar2d_separable`. The redundant `inv_opvar2d` function has been removed.