Create get_eq_opts_2D.m
This function takes a 'opvar2d' or 'dopvar2d' type object on which to impose an equality constraint Qop=Qeop for a decision variable Qeop, and checks what options to use in creating the decision variable Qeop. In particular, if certain parameters of Qop are all zero, these same parameters in Qeop must be zero, so the output settings will reflect this through the field `eq_opts.exclude`. In addition, if Qop includes only full integrals (no partial integrals) along some direction, the same must be true of Qeop, and the output options will reflect this through the field `eq_opts.sep`.