Add 2D estimator synthesis support.
An executive has been added for constructing an Hinfty optimal estimator for 2D PIEs. An associated `getObserver` function has also been added to take the output of the executive, and build an estimator for the PIE (in particular to build the Luenberger gain operator defining this estimator). The 1D functions have been updated to automatically call the 2D version for 2D systems.