Minor indexing adjustment.
In initializing a PDE structure, the IDs associated to each object were previously stored in the first column of the associated object table (e.g. `PDE.x_tab'). However, this means the index associated to an object may then differ from its ID, which may lead to confusion and errors in e.g. simulation. The function has now been updated such that the initialized structure has the index rather than the ID for each object stored in the first column of the object table, keeping track of the ID only through the explicit ID field, e.g. `PDE.x{1}.ID'. The reorder function has also been updated to return the new order of the indices, rather than the new order of the IDs.