Update convert_PIETOOLS_PDE_2D.m
There was an issue where a PIE representation could not be found for PDEs on domains with larger limit values (e.g. [10,12]), even though a PIE representation could be obtained for the same PDE on a domain of same the same size but with smaller limits ([0,2]). The cause seems to be numerical, in that the inverse computed in order to construct the PIE representation is only accurate up to certain tolerance, and this accuracy decreases with the size of the domain. For now, the required tolerance in the accuracy of the inverse has been increased to be more flexible, but for sufficiently large domains this still does not fix the issue. Perhaps a rescaling would have to be done to avoid this.