Bugfix PDE display.
A bug has been fixed in the case that no identifier has been specified for certain state components or inputs. Previously, this led to an error as the function relies on this identifier for display. Since the non-command line input format does not specify such identifiers, the functions have been changed to use the object index (rather than identifier) for display in this case. In addition, support has been added for vector-valued differential operators, and for evaluating the state components at multiple positions within a single term declaration, displaying the result using square-bracket matrix notation.