Update poslpivar_2d.m
Add support for declaring the degree `d` of the polynomials defining the operator using a scalar value (rather than cell). However, perhaps the input format needs to be changed further for more intuitive declaration of the degree, as well as the size of the object and the variables of the object (which are currently extracted from the `prog` structure, but this `prog` structure may not store them in the right order). Also, the default operator without any specified settings does not impose and separability constraints, making for an expensive optimization program even for small degrees.