Add support for boundary forcing.
The 2D PDE to PIE conversion function has been updated to add support for distributed forcing along the boundary. The support is still limited, in that boundary forcing may only be performed for either Dirichlet or Neumann BCs, or some combination of the two (Robin BCs). Also, it is crucial that the forcing at the boundary does not yield conflicting boundary conditions, as otherwise the PIE representation is invalid (e.g. u(y=0)=w1(x) and u(x=0)=w2(y) is only supported if w1(0)=w2(0)). Note that inputs at the boundary may also appear as derivatives in the PIE representation (just as the PDE state gets converted to a fundamental state in the PIE), so this has to be taken into account for e.g. L2-gain analysis and simulation.