Bugfix 2D converter.
In order to convert a PDE to a PIE, the converter first reorders the state components, inputs, and outputs, based on the number of variables they depend on and their order of differentiability. However, when calling `convert` for a 2D system, this reordering was previously performed twice. Although this is not a problem, the new order of the variables was no longer properly stored in the resulting output PIE structure. This has been resolved, though a better solution should really be implemented once the pull request has been merged with master, as other adjustments to the `reorder_comps` function have already been implemented there.