Add further [matrix;opvar] support.
The concatenation functions now also support concatenations `[P;M]` and `[P,M]` for `P` an opvar and `M` a matrix, if the operator defined by `P` is in fact just a matrix as well. This is useful for e.g. constructing the operator in the KYP lemma, where we may perform concatenation `[P1,M1,P2]` for `P1` an opvar object that acts just as a matrix.