Introduce 'clean_opvar'.
This function removes terms from parameters of an 'opvar' or 'dopvar' object that are defined by coefficients below some tolerance. This can be useful as e.g. integration of 'polynomial' class objects tends to introduce terms with very small coefficients due to finite numerical accuracy, so the `clean_opvar` function can be used to get rid of these terms. The function was already introduced for `opvar2d` objects, for which this issue is more prevalent.