Add `diff` and `subs` for opvar.
These functions allow compositions of PI operators with differential and Delta operators to be computed. In particular, for an 'opvar' object 'P', The function `diff` computes the operator 'dP'` such that 'd/dx*P*u=dP*[u;u_{x}]', or, if possible and requested, 'd/dx*P*u=dP*u'. The function `subs` takes a position argument 'a', and computes 'DP' such that '(P*u)|_{x=a}=DP*[u(a);u]', or, if possible and requested, '(P*u)|_{x=a}=DP*u'. These functions should replace `diff_opvar` and `delta_opvar` (the latter of which does not seem to exist anyway).