Combine 1D and 2D converters.
The function `convert_PIETOOLS_PDE` now acts as converter for both 1D and 2D PDEs to PIEs. This conversion is supported for infinite-dimensional inputs on the interior of the domain, and limited infinite-dimensional inputs at the boundary in the 2D case. The old 2D converter has been removed, and the 1D converter has been moved to to the legacy folder.