Update opvar subsref.
The different `subsref` functions for 'opvar' and 'opvar2d' objects now allow the spatial variables to be extracted as e.g. `Pop.vars`, so that `Pop.vars(:,1)=Pop.var1` and `Pop.vars(:,2)=Pop.var2`. This allows an LPI program structure on the same variables and domain as an operator `Top` to be specified as `prog=lpiprogram(Top.vars,Top.I)`.