Create mat2opvar.m
This function takes a matrix or polynomial, and converts it to an object of type 'opvar', or 'opvar2d', depending on the number of specified dimensions and variables. The function assumes the matrix to represent a multiplier operator, and will therefore only set parameters in the 'opvar' structure corresponding to multiplier operators, throwing an error if some nonzero block in the matrix must represent a kernel of an integral operator.