Update executives.
The function `mat2opvar` is now used to convert the identity matrices used in the negativity constraints in the KYP lemma to PI operators. This ensures that concatenation can be performed without ambiguity, though some cases may still require us to revert to the old `horzcat` and `vertcat` functions (for example if the disturbance is infinite-dimensional, and the output finite-dimensional). In the 2D executives, issues with these cases are avoided by using e.g. `horzcat_legacy`.