Update lpiprogram.m
The function now requires only primary spatial variables to be specified as the first argument, rather than also the dummy variables used for integration. If only a column or row vector of variables is passed as the first argument, the function assumes only primary spatial variables are specified, and generates associated dummy variables itself by taking the spatial variable names and adding '_dum'. Note that there is a bit of ambiguity if variables are passed as a 1x2 row array `vartable=[x,y]`, in which case it is not obvious whether `x` and `y` are separate spatial variables, or if `x` is a spatial variable and `y` is a dummy variable. In this setting, the function relies on the domain `dom` to be properly specified, so that if `dom` has 1 row it assumes `y` is a dummy variable, and if `dom` has two rows it assumes `y` is a primary spatial variable on domain `dom(2,:)`.