Bugfix executives.
There seems to be an issue with addition of 'opvar' objects and matrices, as a result of which, adding a small positive matrix to positive semi-definite operators in the executives returned a zero operator. This has been resolved by converting the matrices to operators before addition, but the underlying issue with addition of opvar and non-opvar objects has not yet been solved.