Simplify demos 1-6.
The first 6 demos have been simplified to an extent. Further simplification may be possible; especially the plotting of simulated results is still quite messy. The biggest change is perhaps to demo 3, where now Finsler's lemma is used to enforce the second boundary condition, rather than requiring the state to be second-order differentiable. The obtained result is more accurate, and avoids the need for declaring a PDE system which is otherwise not really used. That being said, it does require the user to be familiar with Finsler's lemma, and adapting the test to different boundary conditions may be complicated.
No updates have yet been made to demo 7.