Create DEMO9_L2_gain_analysis.m
A demo illustrating how a 2D PDE can be declared, and the L2 gain of the system can be tested by solving an LPI. However, it seems `lpi_ineq2d` is not able to enforce the inequality constraint with sufficient freedom, as a result of which the program is infeasible. This will need to be fixed.