Update Demo 8.
The demo has been updated to account for the "lpi_programming" update to the master branch. In particular, calls to e.g. `sosprogram` have been replaced with corresponding calls to `lpiprogram`, as the old version of the demo would no longer work in the master branch. Other than that, the demo remains mostly the same. Potential further changes to the demo (e.g. adding version history) can be made in the "simplify_demos" branch.