Update opvar2d multiplication.
The `mtimes` functions now check whether the primary and dummy variables of the operators to compose indeed match, to avoid errors. If the variables do not match, the primary variables of the left operator are used, and the dummy variables of the right operator. However, in the case of `dopvar2d`-`opvar2d` multiplication, the variables of the `dopvar2d` object are assumed to be correct, since those variables will appear in some LPI program.