Update lpiprogram.m
The program now takes dummy variables as a separate, optional, second argument. For backward compatibility, it is still possible to include dummy variables in the first argument as well, specifying primary variables in the first column of `vartab` and dummy variables in the second column, if the number of rows of `vartab` matches the number of specified domains. If this feature is removed, the demos and executives will need to be changed to specify the dummy variables separately.