Update PI operator discretization.
The function `PIESIM_3PI2Mat_cheb_opint_discretize` has been updated to take as additional arguments the variables defining the PI operator. This is to avoid substitution of erroneous variables in discretizing the operator. Any calls to this function have also been updated, namely in  `PIESIM_3PI2Mat_cheb`.
Of course, with the update to PIESIM in the 2D simulation branch, these updates may become obsolete, or may need to be reperformed.