Update default vars in 'sys' structure.
The functions for the command line input format have been updated to replace the default dummy variable 'theta' with 's_dum'. Note that the primary spatial variable is still just 's' rather than 's1', as otherwise users will have to use 's1' as well when declaring their PDEs, which conflicts with the demos and manual. Of course, with the ND parsing update, these changes may become obsolete, but for now this should resolve some errors when declaring PDEs.