Update opvar addition.
Addition of objects `Pop` of type 'opvar' ('dopvar', 'opvar2d') and objects `A` of type 'double' or 'polynomial' is now supported more generally, treating `A` as a multiplier operator. This only works if the dimensions of `A` match those of `Pop`, and the block of elements in `A` that would be added to the parameters `Pop.Q1` (representing an integral term) are zero, as we assume that `A` cannot represent an integral operator. This check is performed in the function `mat2opvar`. The purpose of this update is to allows strict positivity of a semidefinite PI operator `Pop` to be enforced as `Pop=Pop+eppos*eye(size(Pop))`, without the need for users to convert `eye(Pop)` to a PI operator themselves.

In addition, the `plus` functions have been updated to make sure that the spatial and dummy variables defining two 'opvar' objects `Pop1` and `Pop2` in a sum `Pop1+Pop2` match, as otherwise, the resulting operator is not well-defined. The functions `eq` have been adjusted to account for this change.