Update lpidecvar.m
The function can now be used to generate new decision variables of a desired size, rather than just to add (already generated) individual decision variables to the optimization program. This is to make the use of this function more similar to that of `poslpivar` and `lpivar`, and avoid the need for users to declare decision variables separately themselves using `dpvar`.