Update opvar2d concatenation.
The concatenation functions have been updated to allow concatenation of 'opvar2d' and 'dopvar2d' objects with matrices or matrix-valued polynomials. Support is still very limited, as the matrix is assumed to either map to R (vertical concatenation) or from R (horizontal concatenation). The main purpose is for e.g. the L2 gain LPI, where the decision variable gamma to minimize need now no longer be converted to an opvar2d object manually.