Update lpi_ineq_2d.m
The function now checks whether the input operator is separable, and whether any parameters can be excluded, no longer allowing the user to specify such options. The user can only specify whether to add a psatz term, and below what tolerance coefficients in the operator are taken to be zero.