Create pielft.m
The function takes the linear fractional transformation of two PIEs, setting the observed outputs of the first PIE equal to the controlled inputs of the second PIE, and vice versa, so that the closed-loop system has no more observed outputs and controlled inputs. Future updates may expand functionality to e.g. specify which outputs map to which inputs.