Update pde_var representation.
A loose PDE variable (i.e. a state, input, or output) was previously represented as 'pde_struct' object without explicitly declaring any term involving this variable. For example, an output "y" was represented as a 'pde_struct' object `y` by specifying only the values of the fields of `y.y{1}`, such as `y.y{1}.size` and `y.y{1}.vars`. However, this is also how we would represent a zero equation "y=0", leaving ambiguity as to whether the object `y` referred to just an output variable or an actual (though trivial) equation. This also resulted in an error in `vertcat`, where zero equations are treated as boundary conditions. To resolve this, loose PDE variables are now assigned an actual term, e.g. "1*y", declared as a single element in the field `y.free{1}.term{1}`, just as we would declare more general terms. This way, we can distinguish a term defined as just a single variable from an actual zero equation for this variable. Consequently, the function `var2term` that was previously used to convert a loose PDE variable (e.g. "y") to such a term (e.g. "1*y") has become redundant, and a number of functions have been updated to account for this fact.
In addition, since (before initialization) boundary conditions are expressed in terms of state variables that may not have been assigned an equation yet, the update would have interpreted the empty terms for the state variable as a differential equation, even if such an equation had not actually been declared. This would cause an issue in display, displaying both the boundary condition and a non-existent differential equation. This has been resolved by adding a field "is_zero" as e.g. `PDE.x{1}.is_zero=true`, to indicate whether the corresponding variable has been assigned a zero equation, or whether it is just empty for use in e.g. the boundary conditions. This field is removed again during initialization.