Update to pde_var subsasgn.
For state variables, the desired order of differentiability can now be explicitly declared through the field `diff`, setting e.g. `x.diff=[2;1]` to indicate that variable `x` is twice differentiable with respect to its first spatial variable, and once differentiable with respect to its second one. Alternatively, this order of differentiability can also be passed as final argument to the function `pde_var`.