Update command line input examples.
The examples have been updated to use the `pde_var` function rather than `sys` and `state`. For the function `stability_3`, however, the example has been implemented slightly differently, as the approach used in the `sys` structure fails for the `pde_var` implementation. This is a shortcoming that should be addressed in a future release.