Update eq.m
The function previously did not allow rows of vector-valued terms to be equated unless there were equal numbers of rows on the left- and right-hand side, and the vector size of each row on the left matched that on the right. With this update, more general equations can be enforced as well, so long as the sum of the vector-sizes of the rows of terms on the left-hand side matches that of the terms on the right-hand side. This is achieved by combining all rows on the right-hand side into a single row of higher-dimensional vector-valued terms, and similarly for the left-hand side, and then equating those terms.