Update Hinf_gain to non-coercive version.
I updated the LPI executive to use the non-coercive approach, declaring a positive operator Rop(=Top'*Pop*Top)>=0 and indefinite operator Qop(=Pop*Top), and enforcing equality Top'*Qop-Rop==0. To generate the indefinite operator Qop, I created a function `get_lpivar_degs` that takes as input the operator Rop, and returns an array of degrees to be passed to `lpivar` when declaring Qop. The function is not very smart, just checking maximal degrees of Rop, but seems to do okay.
The coercive version of the LPI is still available, renamed to reflect its coercive nature.