Update initialize.m
Fixed a bug, allowing a location to be specified for terms involving inputs as well. This location is initially removed in the subroutine `get_diff_tab`, but may be reintroduced when taking an integral of the input, throwing an error when this field is encountered again later despite it not being expected. This has been resolved by explicitly checking whether a location has been specified even for terms with inputs, though a more extensive overhaul of which fields are allowed for input signals and which aren't may need to be performed at some point.