Bugfix separable opvar2d inversion.
There were two issues with inversion, that have been resolved. One had to do with trying to evaluate `Bp/Ap` for a polynomial class object `Bp`, which is not defined. The other had to do with expanding a polynomial `R(s)` into`R(s)=Z(s)'*C*I`, where the dimensions of the identity operator `I` were not properly set.