Added PDDE example
Created non-coercive H2 norm executives in 2D, but couldn't run it well in the PDDE example. This needs to be further tested. For now the H2 norm executive uses the coercive versions in 2D cases. Additionally,there was a bug in running the feasebility tests of the coercive scripts. The input gam should be squared in the inquality gam^2>trace().