Update opvar-matrix multiplication.
Multiplication of opvar objects with non-opvar objects is now supported only if the opvar object maps to (for matrix x opvar) or from (for opvar x matrix) a single space, e.g. R or L2 (not RxL2). In that case, the matrix is assumed to act as multiplier operator on that space. If the operator is more general, an error is thrown warning the user of the ambiguity.