Update batch converter.
The conversion function for PDEs in batch format now also allows the spatial variables appearing in the system to be declared in `PDE.vars`. The initialization function has been updated to check the parameters for possible spatial variables, and warn the user if it is unclear which variable acts as spatial variable.