Update PIESIM_plot_solution.m
There was an issue when plotting solutions involving large numbers of state variables, that insufficient plot colors were specified for all states. The function now checks whether the number of state variables exceeds the number of colors, in which case it starts using the Matlab default colors. However, if this does occur, we have a lot of subplots in one figure, which is still rather messy.