Update diff_opvar.m
The function now checks whether the parameters defining the input 'opvar' object are indeed of type 'polynomial', avoiding an error in which a derivative was attempted to be taken of objects of type 'double'. Also, the variables of the output object `dP` are now set equal to those of the input object `P`.