Update opvar inversion.
The `inv_opvar_old` function has been updated to support inversion of non-self-adjoint operators, as well as to fix a bug in case parameters (e.g. `P.R.R0`) are not polynomial, causing problems with empty degmats. Also, the `inv` function for opvars has been updated to allow a call to `inv` without specifying a tolerance, given that this is also supported by the `inv_opvar` routine, for which `inv` is a shell.