Update default SDP solver.
Instead of declaring 'sedumi' as the SDP solver to use for solving LPI programs, the different scripts and demos now no longer declare any particular solver to use. As a result, 'sossolve' (as called by 'lpisolve') will check for any SDP solvers on the Matlab path, and use the first SDP solver it encounters, throwing an error if no solver is found.