Update opvar subsagn.
An @opvar/subsasgn function has been created, so that elements of 'opvar' objects can be set (e.g. `Pop([1,3],[2,4])=  [1,0;0,3]`). If the right-hand side of the imposed equality is not of type 'opvar', the function will try to convert it to an 'opvar' object, or throw an error if this fails. Naturally, the dimensions of the object on the right-hand side should match those of left-hand side. A function `end` has also been created, so that we can extract and declare elements of an 'opvar' object as e.g. `Pop(1:end,:)`. Also fixed a bug with the @opvar2d/subsasgn function.