Update `expand_tderivatives`
This function takes a PDE involving higher-order temporal derivatives of certain PDE states, and introduces auxiliary state variables to model these higher-order derivatives using only a first-order in time PDE. Previously, these auxiliary variables were not required to be differentiable in space, and no boundary conditions were imposed. Now, the function has been updated to check what boundary conditions are imposed on the PDE state variables, and impose those same boundary conditions on the associated auxiliary variables. Note that this may introduce ambiguity when state variables are coupled in the boundary conditions, e.g. if we have `x1(s=0)=x2(s=0)`, but `x1` is only first-order differentiable in time, and `x2` is second-order differentiable. The function will display a warning in this case. The old function has been renamed, so that it may still be used for comparison (if desired).