Add 'misc' field to pie structure.
The `pie_struct` object class has been updated to add a field `misc`, in which additional objects/data relevant to the PIE can be added to the user. The display function has been updated to display the contents of this field as well. However, `horzcat` and `vertcat` still need to be updated to account for this added field.